Move ResponseCompleteException to common
